# [xy] Speed up charts.