# Summary
<!-- Brief summary of what your code does -->
Speed up multiple charts.
Benchmark: for a dataset with 10 million rows
* Pie chart: 1s -> 0.6s
* Unique values: 2.9s -> 1s
* Most frequent values: 3.1s -> 2.4s

# Tests
<!-- How did you test your change? -->
tested locally

cc:
<!-- Optionally mention someone to let them know about this pull request -->